receipt-gen-ai/pyproject.toml
[project]
name = "receipt-gen-ai"
version = "0.1.0"
description = "Fake receipt generator (JSON + realistic image)"
authors = [{ name="Pierre Ribeiro", email="pierre@ribeiro.com.pt" }]
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"

dependencies = [
  "openai>=1.25.0",
  "anthropic>=0.25.0",
  "faker>=24.0.0",
  "pyyaml>=6.0.1",
  "python-dotenv>=1.0.1",
  "typer[all]>=0.9.0",
  "jinja2>=3.1.3",
  "rich>=13.7.0"
]

[project.optional-dependencies]
dev = ["pytest>=8.2.0"]

[project.scripts]
receipt-gen-ai = "core.cli:app"

[build-system]
requires = ["setuptools>=67", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["core"]
package-dir = {"" = "src"}
